pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cc-framework"
version = "0.2.0"
description = "Composability Coefficient (CC) framework for measuring guardrail interactions"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Pranav Bhave" }]
keywords = ["AI safety", "guardrails", "security", "evaluation", "statistics", "composability"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Security",
  "Topic :: Scientific/Engineering",
]

# CORE = minimal install required to run theory + core data layer.
# Heavy scientific stack moved to extras so users/reviewers can opt in.
dependencies = [
  "numpy>=1.26",
  "pydantic>=2.0",
  "pyyaml>=6.0.1",
  "jsonlines>=3.0",
  "cryptography>=41.0",
  "blake3>=0.3",
]

[project.optional-dependencies]
# Scientific stack used by theory + inference utilities (optional by design)
stats = [
  "scipy>=1.11",
  "statsmodels>=0.14",
]

# Visualization stack
viz = [
  "matplotlib>=3.8",
  "seaborn>=0.12",
  "plotly>=5.0",
]

# Dataframe convenience
data = [
  "pandas>=2.2",
]

# ML helpers (if/when you use sklearn in adapters or evaluation pipelines)
ml = [
  "scikit-learn>=1.3",
]

# Notebook UX
notebooks = [
  "ipykernel",
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.11",
  "matplotlib>=3.8",
  "pyyaml>=6.0.1",
]

# Serialization/interop
sql = ["sqlalchemy>=2.0"]
avro = ["fastavro>=1.9"]
protobuf = ["protobuf>=5.0", "googleapis-common-protos>=1.63"]

# Docs
docs = ["mkdocs>=1.6", "mkdocs-material>=9.5"]

# Dev tooling
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "hypothesis>=6.90",
  "black>=24.3",
  "ruff>=0.6",
  "isort>=5.13",
  "mypy>=1.10",
  "types-PyYAML>=6.0.12",
]

# Convenience meta-extra: “I’m doing serious work, give me everything”
all = [
  "cc-framework[stats,viz,data,ml,sql,avro,protobuf,docs]",
]

[project.urls]
Homepage = "https://github.com/Cubits11/cc-framework"
Repository = "https://github.com/Cubits11/cc-framework"
Issues = "https://github.com/Cubits11/cc-framework/issues"

# Optional: if you have or plan a CLI entry point, wire it here.
# (Remove if you don't have cc/cli.py:main)
# [project.scripts]
# cc = "cc.cli:main"

# -----------------------
# setuptools src-layout
# -----------------------
[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["cc*"]
exclude = [
  "tests*",
  "benchmarks*",
  "experiments*",
]

[tool.setuptools.package-data]
"cc.adapters" = ["nemo_configs/minimal/*"]

# -----------------------
# pytest
# -----------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -ra --maxfail=1 --strict-config --strict-markers"
markers = [
  "experiment: gated experiments producing artifacts",
  "perf: gated performance benchmarks",
  "slow: slow tests that are not run by default",
]
filterwarnings = [
  "error::DeprecationWarning:cc\\..*",
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning",
  "ignore:.*distutils.*:DeprecationWarning",
]

# -----------------------
# mypy
# -----------------------
[tool.mypy]
python_version = "3.10"
mypy_path = ["src"]
packages = ["cc"]
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true
warn_return_any = true
no_implicit_optional = true
show_error_codes = true
pretty = true

# A “research codebase” benefit: tighter type discipline catches logical drift.
# If this is too aggressive initially, set to false and ratchet up module-by-module.
strict = true

exclude = [
  "tests/",
  "benchmarks/",
  "experiments/",
]

# -----------------------
# ruff (lint + format)
# -----------------------
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["build", "dist", ".venv", ".git", ".hypothesis", ".pytest_cache"]

[tool.ruff.lint]
# Strong but sane default set for a scientific/security codebase.
select = ["E", "F", "W", "I", "UP", "B", "SIM", "C4", "PIE", "RUF"]

# We are explicitly not spending cycles on typing deprecation churn.
# (Your codebase can modernize these in a dedicated refactor PR later.)
ignore = [
  "E501",  # handled by formatter/line-length policy
  "UP035", # typing.List/Dict/Tuple/Type/Set deprecation warnings (PEP 585 churn)
  "UP006", # dict/list/tuple builtins vs typing.* generics (paired with UP035)
  "UP007", # X | Y vs Union[X, Y] (paired with UP035)
]

[tool.ruff.lint.isort]
known-first-party = ["cc"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
  "B011",
  "SIM117",
]
"experiments/**.py" = [
  "B011",
  "SIM117",
]

[tool.ruff.format]
quote-style = "double"

# -----------------------
# black (optional; ruff can format too, but keeping black helps external contributors)
# -----------------------
[tool.black]
line-length = 100
target-version = ["py310"]
exclude = '''
(
  /(
      \.venv
    | build
    | dist
    | \.git
    | \.hypothesis
    | \.pytest_cache
  )/
)
'''

# -----------------------
# isort (optional; ruff can sort too, but keeping isort helps external contributors)
# -----------------------
[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "tests", "experiments"]

# -----------------------
# coverage
# -----------------------
[tool.coverage.run]
source = ["src/cc"]
branch = true
parallel = true
relative_files = true
omit = [
  "tests/*",
  "benchmarks/*",
  "experiments/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = true
exclude_lines = [
  "if __name__ == .__main__.:",
  "pragma: no cover",
]